little fix

+++ srcs/config/Path.cpp
@@ -6,7 +6,7 @@
 /*   By: jeada-si <jeada-si@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/01/06 13:46:37 by jeada-si          #+#    #+#             */
-/*   Updated: 2025/01/08 10:56:22 by jeada-si         ###   ########.fr       */
+/*   Updated: 2025/01/08 12:02:10 by jeada-si         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -169,7 +169,7 @@ std::ostream&	operator<<(std::ostream & out, const Path & path)
 
 bool	Path::throughParent() const
 {
-	size_t	i = 0;
+	long	i = 0;
 
 	for (t_strVec::const_iterator it = _path.begin();
 		it != _path.end(); it++)
